Move hyper runtime wrappers to separate module
